I made some changes of the game. All changes work well.
One team member is added after two chest boxes are opened. There is another person wants to join in the team and I put "Yes" and "No" answers to choose. If you choose "No", he will keep saying he wants to join in. If you choose "Yes", he will be a new member. A door appears which can be chosen to remove after he joins in.